Move text into primitives
