+++ library/Rattletrap/AttributeValue/PrivateMatchSettings.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.PrivateMatchSettings where
 
-import Rattletrap.Text
+import Rattletrap.Primitive.Text
 import Rattletrap.Word32
 
 import qualified Data.Binary.Bits.Get as BinaryBit
