+++ library/Rattletrap/AttributeValue/Reservation.hs
@@ -2,7 +2,7 @@ module Rattletrap.AttributeValue.Reservation where
 
 import Rattletrap.AttributeValue.UniqueId
 import Rattletrap.Primitive.CompressedWord
-import Rattletrap.Text
+import Rattletrap.Primitive.Text
 import Rattletrap.Word8
 
 import qualified Data.Binary.Bits.Get as BinaryBit
