+++ library/Rattletrap/Property.hs
@@ -1,7 +1,7 @@
 module Rattletrap.Property where
 
 import Rattletrap.PropertyValue
-import Rattletrap.Text
+import Rattletrap.Primitive.Text
 import Rattletrap.Word64
 
 import qualified Data.Binary as Binary
